Refactor SSR test helpers
