+++ packages/@headlessui-react/src/components/tabs/tabs.ssr.test.tsx
@@ -1,9 +1,6 @@
-import { RenderResult } from '@testing-library/react'
-import { render, RenderOptions } from '@testing-library/react'
-import React, { ReactElement } from 'react'
-import { renderToString } from 'react-dom/server'
+import React from 'react'
 import { Tab } from './tabs'
-import { env } from '../../utils/env'
+import { renderSSR, renderHydrate } from '../../test-utils/ssr'
 
 beforeAll(() => {
   jest.spyOn(window, 'requestAnimationFrame').mockImplementation(setImmediate as any)
@@ -31,7 +28,7 @@ function Example({ defaultIndex = 0 }) {
 describe('Rendering', () => {
   describe('SSR', () => {
     it('should be possible to server side render the first Tab and Panel', async () => {
-      let { contents } = await serverRender(<Example />)
+      let { contents } = await renderSSR(<Example />)
 
       expect(contents).toContain(`Content 1`)
       expect(contents).not.toContain(`Content 2`)
@@ -39,7 +36,7 @@ describe('Rendering', () => {
     })
 
     it('should be possible to server side render the defaultIndex Tab and Panel', async () => {
-      let { contents } = await serverRender(<Example defaultIndex={1} />)
+      let { contents } = await renderSSR(<Example defaultIndex={1} />)
 
       expect(contents).not.toContain(`Content 1`)
       expect(contents).toContain(`Content 2`)
@@ -51,7 +48,7 @@ describe('Rendering', () => {
   // Skipping for now
   xdescribe('Hydration', () => {
     it('should be possible to server side render the first Tab and Panel', async () => {
-      const { contents } = await hydrateRender(<Example />)
+      const { contents } = await renderHydrate(<Example />)
 
       expect(contents).toContain(`Content 1`)
       expect(contents).not.toContain(`Content 2`)
@@ -59,7 +56,7 @@ describe('Rendering', () => {
     })
 
     it('should be possible to server side render the defaultIndex Tab and Panel', async () => {
-      const { contents } = await hydrateRender(<Example defaultIndex={1} />)
+      const { contents } = await renderHydrate(<Example defaultIndex={1} />)
 
       expect(contents).not.toContain(`Content 1`)
       expect(contents).toContain(`Content 2`)
@@ -67,68 +64,3 @@ describe('Rendering', () => {
     })
   })
 })
-
-type ServerRenderOptions = Omit<RenderOptions, 'queries'> & {
-  strict?: boolean
-}
-
-interface ServerRenderResult {
-  type: 'ssr' | 'hydrate'
-  contents: string
-  result: RenderResult
-  hydrate: () => Promise<ServerRenderResult>
-}
-
-async function serverRender(
-  ui: ReactElement,
-  options: ServerRenderOptions = {}
-): Promise<ServerRenderResult> {
-  let container = document.createElement('div')
-  document.body.appendChild(container)
-  options = { ...options, container }
-
-  if (options.strict) {
-    options = {
-      ...options,
-      wrapper({ children }) {
-        return <React.StrictMode>{children}</React.StrictMode>
-      },
-    }
-  }
-
-  env.set('server')
-  let contents = renderToString(ui)
-  let result = render(<div dangerouslySetInnerHTML={{ __html: contents }} />, options)
-
-  async function hydrate(): Promise<ServerRenderResult> {
-    // This hack-ish way of unmounting the server rendered content is necessary
-    // otherwise we won't actually end up testing the hydration code path properly.
-    // Probably because React hangs on to internal references on the DOM nodes
-    result.unmount()
-    container.innerHTML = contents
-
-    env.set('client')
-    let newResult = render(ui, {
-      ...options,
-      hydrate: true,
-    })
-
-    return {
-      type: 'hydrate',
-      contents: container.innerHTML,
-      result: newResult,
-      hydrate,
-    }
-  }
-
-  return {
-    type: 'ssr',
-    contents,
-    result,
-    hydrate,
-  }
-}
-
-async function hydrateRender(el: ReactElement, options: ServerRenderOptions = {}) {
-  return serverRender(el, options).then((r) => r.hydrate())
-}
